Marina/requirements.txt
astroid==2.3.3
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pylint==2.4.3
pyOpenSSL==19.0.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
